requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
markdownify==0.11.6
PyYAML==6.0.1
schedule==1.2.0
python-dateutil==2.8.2
